Restructuring the data folder
Inspired by the way the AzerothCore community works, we modified the directory where the SQL files are stored and also added the SQL pending files folders for future contributors. With these scripts, you can create files to upload the SQL data that you want to contribute to the project. So that later, with the rename script, you can upload and rename the file, once the pull request is approved.